@@ -15,12 +15,22 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.3.3",
-    "@babel/preset-env": "^7.3.1",
+    "@babel/core": "^7.8.4",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-proposal-decorators": "^7.8.3",
+    "@babel/plugin-proposal-export-default-from": "^7.8.3",
+    "@babel/plugin-proposal-export-namespace-from": "^7.8.3",
+    "@babel/plugin-transform-runtime": "^7.8.3",
+    "@babel/preset-env": "^7.8.4",
+    "@babel/preset-react": "^7.8.3",
+    "@babel/preset-typescript": "^7.8.3",
+    "@types/classnames": "^2.2.9",
+    "@types/react": "^16.9.19",
+    "@types/react-dom": "^16.9.5",
     "babel": "^6.23.0",
     "babel-core": "^6.26.3",
-    "babel-eslint": "^10.0.1",
-    "babel-loader": "^7.1.5",
+    "babel-eslint": "^10.0.3",
+    "babel-loader": "^8.0.6",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.7.0",
@@ -39,10 +49,13 @@
     "file-loader": "^1.1.11",
     "html-webpack-plugin": "^3.2.0",
     "ip": "^1.1.5",
-    "node-sass": "^4.12.0",
+    "node-sass": "^4.13.1",
     "react-hot-loader": "^4.1.2",
     "sass-loader": "^7.0.1",
+    "source-map-loader": "^0.2.4",
     "style-loader": "^0.21.0",
+    "ts-loader": "^6.2.1",
+    "typescript": "^3.7.5",
     "url-loader": "^1.0.1",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.2.3",
@@ -72,16 +85,17 @@
     "monaco-editor-webpack-plugin": "^1.8.0",
     "pretty-seconds": "^0.2.1",
     "query-string": "^6.1.0",
-    "react": "^16.6.3",
+    "react": "^16.12.0",
     "react-autosize-textarea": "^6.0.0",
     "react-dnd": "^7.0.2",
     "react-dnd-html5-backend": "^7.0.2",
-    "react-dom": "^16.8.3",
+    "react-dom": "^16.12.0",
     "react-fontawesome": "^1.6.1",
     "react-json-editor-ajrm": "^2.5.9",
     "react-json-pretty": "^2.0.1",
     "react-number-format": "^4.0.7",
     "react-redux": "^5.0.7",
+    "react-resize-detector": "^4.2.1",
     "react-router-redux": "^5.0.0-alpha.9",
     "react-table": "^6.8.6",
     "react-tooltip": "^3.9.0",
