@@ -1,6 +1,6 @@
-export ReactTableComponent from './ReactTable'
-export TableHeader from './TableHeader'
-export Pagination from './Pagination'
-export SearchResults from './SearchResults'
-export NoMoreResults from './NoMoreResults'
-export LoadMoreResults from './LoadMoreResults'
+export { default as ReactTableComponent} from './ReactTable';
+export {default as TableHeader} from './TableHeader';
+export {default as Pagination} from './Pagination';
+export { default as SearchResults} from './SearchResults';
+export { default as NoMoreResults} from './NoMoreResults';
+export { default as LoadMoreResults} from "./LoadMoreResults";
